Modify 1.7 install instructions for headless
Most importantly, this instructs users to set the `_SEARCH_ENABLED`-suffixed
env vars to `'false'` if they want to run Archivematica "headless", i.e., with
Elasticsearch not installed. This commit also includes some formatting changes.

Also removes the section entitled `Upgrade from Archivematica 1.8.0 to 1.8`
because it did not seem to make sense. This section may need to be restored and
re-written.