Modify "Upgrade in indexless mode" section
The "Upgrade in indexless mode" section now contains instructions for
turning indexing off in AM as well as stopping and disabling a running
Elasticsearch service.

Also removes the package URL Sphinx substitutions in conf.py because the
formatting of the `parsed-literal` directives that contained these was
not working.

Also made formatting modification so that code blocks appear correctly
indented (e.g., within <li>s) in the built HTML.