Simplify preservation planning page, add FPR docs, remove FPR-server
Condenses and simplifies the preservation planning page by removing
explanatory text. Removes references to the FPR Server and updating
rules, while adding content about how new rules/formats are added
(e.g. with new releases). Adds content from the fpr-docs so that
Preservation Planning and FPR page is all on one page.

Note that this commit does not fully address #86 because this commit
is not backwards-compatible- we need a different commit for versions
1.6.1 and lower.