Documentation for resolving decision points
This addition to the documentation details how to use the MCPClient to resolve
hanging decision points, which are caused by users removing a transfer or SIP
from the dashboard before resolving a decision that is awaiting user input.

It also fixes up the links to use inline literals and fixes some code-block
formatting.

@mamedin, can you double-check the code snippets and provide corrections if
needed? I didn't have an Archivematica instance to work with to replicate
the steps so I took an educated guess - I'm sure there are errors.

Connected to archivematica/Issues/#67