Update Administration Tab pages
This change was a bit of a rabbit hole - I started out updating the config
information for ArchivesSpace DIP uploads and then realized that a good clean-
up was in order.

Note that there are two pages that contain the bulk of the changes:
* User Manual > Administer > Dashboard administration tab
* Admin Manual > Installation and setup > Customizations > Dashboard administration tab

The latter already existed, but was not linked from anywhere except the admin
manual index, which is not normally accessed by users unless they end up there
via the search. I've now linked this page from the Customizations page.

In terms of content, I've attempted to remove administrator-level configuration
information from the user manual and place it in the administrator manual, with
appropriate links between the two.

I've also done a general formatting cleanup to align with our style guide.

This change coincidentally addresses the following:

Connected to archivematica/Issues#170
Connected to archivematica/Issues#168