Add Handle.Net documentation
This change introduces a section on the Integrations page for Handle.Net,
which is the software behind the PID binding feature. I've tried to keep
this change in line with other recent integration documentation updates. I've
added information about configuring the integration in the dashboard and
user manual information on what happens when you select "Yes" for the Bind
PIDs decision point.

Connected to archivematica/Issues#134.

One thing that this change does not address that is suggested in the issue
created by @ross-spencer is the expected output in the AIP METS. If there's an
example out there, I would love to include it.

A few other notes:

* The Handle.Net website uses Handle.Net, Handle System, and Handle Server in
various ways. In an attempt at consistency, I've used Handle.Net as it seems to
be the most common, and is the main term used on the Handle.Net homepage and
on the title page of the documentation.
* As always, I've cleaned up formatting and links where I noticed issues.