Move old FPR document to main docs repo and update
This change adds administrator-level FPR documentation to the main AM docs
repository. Prior to 1.7, we maintained the FPR docs as a branch within
artefactual/archivematica-fpr-admin. I've also substantially updated both the
contents and the formatting of this documentation, though I'm sure there is
still clean-up left to do.

Connected to archivematica/Issues#205. This change will check off the first
to-do item in that issue.