Update Preservation Planning page
As a result of some mis-reading on my part, I thought that the FPR page
still needed some updating. I did that, and then realized that @sromkey
had already done it! However, I didn't want to waste my work so I've applied
some reorganization and reformatting, as well as some content changes, to
this page. It's also being linked now from the Admin Manual as well as the
User Manual.

Connected to archivematica/Issues#205