Update AIC page
This change updates the AIC page for clarity and adds more information about
AIC creation, retrieval, and updating, as well as a new section on the contents
of the AIC package itself.

I've also made some formatting updates to the AIP structure page, as tabbing
was causing some paragraphs to display as block quotes.

Connected to archivematica/Issues#178