Fixes broken links
These errors were probably my fault! Whoops.

This addresses the issues that ablwr discovered in archivematica/Issues#373.
The multiple-machines issue was actually caused by the link to that now-deleted
section being inside the CSV that's used to generate the table shown on the
page. So the Sphinx error message that Travis was displaying was not particularly
clear.