Update bag info page
This change substantially re-writes the bags.rst page, which is the main page
where users can learn about Archivematica's implementation of BagIt, adding
more information about the transfer types, BagIt implementation, and how to add
metadata to the bags.

I've updated the contents as well as screenshots and reformatted it to comply
with our style guide.

Connected to archivematica/Issues#177
Connected to archivematica/Issues#209