Update forensic.rst for clarity and style
This is a general update to the forensic disk images page. I've
clarified the rationale behind some of the processing config options
and added steps for processing compound disk images. I also cleaned up
the formatting and some of the instructions in the steps.

As a result of needing to grab links from other pages, I've also made
some small changes to appraisal.rst, dashboard-admin.rst, and
scaling-archivematica.rst, to comply with the style guidelines.

Connected to archivematica/Issues#668