Update import-metadata.rst (#288)
This started as a small change but turned into a big one - sorry!

I have done a fairly thorough re-write of import-metadata.rst, which is
the page that users are redirected to throughout the documentation when
it comes to adding metadata to a transfer. Changes include:

* Changing the simple/compound descriptive metadata information to more
accurately reflect functionality
* Added information about how non-Dublin Core descriptive metadata is
handled
* Added clearer steps for both of the above
* Added steps for adding rights metadata to a transfer
* Added a METS example for rights
* Added a section with steps on how to add metadata to bags
* Added a small section about including other kinds of metadata

I've also moved the METS examples into snippets that can be used as
inline literals rather than having all that XML directly on the page.
It's something that we started to do for the installation documentation
and it makes working with the documentation pages much nicer.

Finally, I added more CSV examples as required by the above changes.

Minor changes were made to bags.rst and transfer.rst to properly link to
this new documentation.

Resolves archivematica/Issues#694