Update DuraCloud integration page
A general clean-up for the DuraCloud integration page.

After discussion with @ablwr and @dcraze, I have removed the sections
related to using DuraCloud as a transfer source location. Though it is
possible, early tests following the development of the integration
showed that using DuraCloud for this purpose was very slow and possibly
unreliable. Artefactual does not recommend this for Archives Direct
users, so it seemed better to remove the information. It can be added
back if that functionality improves.

This is connected to archivematica/Issues#171