Update DuraCloud integration page (#343)
* Update DuraCloud integration page

A general clean-up for the DuraCloud integration page.

After discussion with @ablwr and @dcraze, I have removed the sections
related to using DuraCloud as a transfer source location. Though it is
possible, early tests following the development of the integration
showed that using DuraCloud for this purpose was very slow and possibly
unreliable. Artefactual does not recommend this for Archives Direct
users, so it seemed better to remove the information. It can be added
back if that functionality improves.

This is connected to archivematica/Issues#171

* Updated to point to SS docs

* Add explicit reference to leading slash

I've added an explicit reference telling users to make sure that they are
copying the DuraCloud space name exactly, ensuring that they're not
including a leading slash.

* Update admin-manual/installation-setup/integrations/duracloud-setup.rst

Co-Authored-By: Ross Spencer <all.along.the.watchtower2001@gmail.com>

* Incorporate feedback and rebase

Co-authored-by: Ross Spencer <all.along.the.watchtower2001+github@gmail.com>