Add section on special bag structures (#354)
I've added a section about alternate configurations for bags. It's a bit
of a punt, as we don't test all of the possible bag configurations, but
I'm not sure how else to communicate that there may be other setup
options out there while also acknowledging that A) Artefactual doesn't
test them all, B) some of them might not work, and C) if they don't work
it doesn't mean it's a bug. So this is what I've landed on.

Connected to archivematica/Issues#669