Adds instrumentation documentation
This commit adds documentation for performance monitoring using
Prometheus and Grafana.

This documentation was written by @cole and minorly edited and
reStructuredText-ified by me.

Ideally, what I would like to do here is merge this first pass (if it is
generally correct) and then expand on it in subsequent PRs, since I
don't actually know how to configure Prometheus and Grafana myself.

Connected to archivematica/Issues#1122

Co-authored-by: Cole <hi@colemaclean.dev>