Update AIP structure page
This page was a bit abandoned - it wasn't linked properly from the index
or any other page, so it had fallen out of date. I've done a
comprehensive update and added a lot more detail about the standard
structure of an AIP. I opted to use tree diagrams instead of screenshots
but maybe they're too fussy? Not sure.