Update sword deposit permissions issue on AtoM2.7
AtoM 2.7 has added a new feature that deletes the DIP directory from the
SWORD deposit after uploading the DIP to AtoM. In order for AtoM to
delete this directory, the AtoM user (www-data or nginx by default) must
have write permissions to this directory in order to delete it.

This commit explains how to change the sword deposit directory
permissions with the setfacl command.